v0.99.0 -- Initial version of recount.bwtool
I couldn't call it recount-bwtool due to R package naming rules. In any case,
the package has a single function called coverage_matrix_bwtool(). It is similar
to recount::coverage_matrix(). The differences are that:

* it uses bwtool (hence the name), which is much faster than derfinder (once
you have installed bwtool)
* it returns a RangedSummarizeExperiment object instead of a matrix
* the counts are not scaled, but can easily be scaled using recount::scale_counts()

Note that the function will automatically detect if you are working at JHPCE
as part of "leekgroup" or if you are working at SciServer. In those scenarios,
you won't have to download the bigwig files, which will make everything much
faster.